move some startup steps from OHazelcast to ODistributedAbstract plugin
